Move trigger package
