+++ java/src/main/java/mcmp/mc/observability/agent/trigger/mapper/TriggerPolicyMapper.java
@@ -1,7 +1,7 @@
-package mcmp.mc.observability.trigger.mapper;
+package mcmp.mc.observability.agent.trigger.mapper;
 
 import mcmp.mc.observability.agent.common.model.PageableReqBody;
-import mcmp.mc.observability.trigger.model.TriggerPolicyInfo;
+import mcmp.mc.observability.agent.trigger.model.TriggerPolicyInfo;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
