+++ java/src/main/java/mcmp/mc/observability/agent/trigger/service/TriggerTargetService.java
@@ -1,4 +1,4 @@
-package mcmp.mc.observability.trigger.service;
+package mcmp.mc.observability.agent.trigger.service;
 
 import lombok.RequiredArgsConstructor;
 import mcmp.mc.observability.agent.common.exception.ResultCodeException;
@@ -6,10 +6,10 @@ import mcmp.mc.observability.agent.common.model.PageableReqBody;
 import mcmp.mc.observability.agent.common.model.PageableResBody;
 import mcmp.mc.observability.agent.common.model.ResBody;
 import mcmp.mc.observability.agent.monitoring.enums.ResultCode;
-import mcmp.mc.observability.trigger.mapper.TriggerPolicyMapper;
-import mcmp.mc.observability.trigger.mapper.TriggerTargetMapper;
-import mcmp.mc.observability.trigger.model.TriggerPolicyInfo;
-import mcmp.mc.observability.trigger.model.TriggerTargetInfo;
+import mcmp.mc.observability.agent.trigger.mapper.TriggerPolicyMapper;
+import mcmp.mc.observability.agent.trigger.mapper.TriggerTargetMapper;
+import mcmp.mc.observability.agent.trigger.model.TriggerPolicyInfo;
+import mcmp.mc.observability.agent.trigger.model.TriggerTargetInfo;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
