+++ java/src/main/resources/mapper/trigger/history.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="mcmp.mc.observability.trigger.mapper.TriggerHistoryMapper">
+<mapper namespace="mcmp.mc.observability.agent.trigger.mapper.TriggerHistoryMapper">
     <select id="getListCount" resultType="java.lang.Long"></select>
-    <select id="getList" resultType="mcmp.mc.observability.trigger.model.TriggerHistoryInfo"></select>
-    <select id="getDetail" resultType="mcmp.mc.observability.trigger.model.TriggerHistoryInfo"></select>
+    <select id="getList" resultType="mcmp.mc.observability.agent.trigger.model.TriggerHistoryInfo"></select>
+    <select id="getDetail" resultType="mcmp.mc.observability.agent.trigger.model.TriggerHistoryInfo"></select>
 </mapper>
