+++ java/src/main/resources/mapper/trigger/target.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="mcmp.mc.observability.trigger.mapper.TriggerTargetMapper">
+<mapper namespace="mcmp.mc.observability.agent.trigger.mapper.TriggerTargetMapper">
     <select id="getListCount" resultType="java.lang.Long"></select>
 
-    <select id="getList" resultType="mcmp.mc.observability.trigger.model.TriggerTargetInfo"></select>
+    <select id="getList" resultType="mcmp.mc.observability.agent.trigger.model.TriggerTargetInfo"></select>
 
-    <select id="getDetail" resultType="mcmp.mc.observability.trigger.model.TriggerTargetInfo"></select>
+    <select id="getDetail" resultType="mcmp.mc.observability.agent.trigger.model.TriggerTargetInfo"></select>
 </mapper>
